backend/requirements.txt
fastapi
uvicorn
python-dotenv
sqlalchemy
pymysql
passlib[bcrypt]
PyJWT
email-validator
